# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sports World EPS",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api_eps/public/"
        }
    ],
    "paths": {
        "/api/crm2/v1/events/{idUsuario}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "7270c4fed76441dcdd2f340434ee2e58",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "schema": {
                            "type": "integer",
                            "default": 75
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/unassignedClasses/{type}/{idUsuario}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "8af3c6fea0c75b9e47413308e3443a9a",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "default": "trainer"
                        }
                    },
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "schema": {
                            "type": "integer",
                            "default": 75
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/asingClass/trainer": {
            "post": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "fe490c66d94d10d839fd5c33a21c8378",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AssignClassRequest"
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/groupClass/{mail}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "647a158ff7ebd99878a9c8cdd3cee41c",
                "parameters": [
                    {
                        "name": "mail",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "default": "trainer@mail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "requestBodies": {
            "AssignClassRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "instructorId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "fechaClase": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "01-01-2023",
                                        "02-01-2023"
                                    ]
                                },
                                "horaClase": {
                                    "type": "string",
                                    "example": ""
                                },
                                "asistentes": {
                                    "type": "integer",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "secret-key",
                "in": "header"
            }
        }
    }
}